Pivotal 94622114: Back Button Creates New Instances of MainActivity
See https://www.pivotaltracker.com/n/projects/1288738/stories/94622114 for all the gory detail.

MainActivity plays with the requested orientation based on phone/tablet.  However in this case where the user never tries to sign in and the user hits the device back button, the _preWaitDialogOrientation member is never initialized properly.  In fact it takes on a default value of 0 which happens to be the constant for SCREEN_ORIENTATION_LANDSCAPE.  When you change orientations for an Activity, the Android framework will destroy and recreated it.  That is what was happening.

The fix is to simply remember what we've set in onCreate.